Move the classic parser to it's own file
