+++ link-grammar/parse.h
@@ -1,0 +1,3 @@
+
+
+void classic_parse(Sentence, Parse_Options);
